+++ demo/api/src/app.module.ts
@@ -1,5 +1,6 @@
 import {
     AccessLogModule,
+    AzureAiTranslatorModule,
     AzureOpenAiContentGenerationModule,
     BlobStorageModule,
     BLOCKS_MODULE_TRANSFORMER_DEPENDENCIES,
@@ -173,6 +174,7 @@ export class AppModule {
                 PredefinedPageModule,
                 CronJobsModule,
                 ProductsModule,
+                ...(config.azureAiTranslator ? [AzureAiTranslatorModule.register(config.azureAiTranslator)] : []),
                 AccessLogModule.forRoot({
                     shouldLogRequest: ({ user }) => {
                         // Ignore system user