+++ packages/admin/admin-theme/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @comet/admin-theme
 
+## 6.16.0
+
+### Patch Changes
+
+-   @comet/admin-icons@6.16.0
+
+## 6.15.1
+
+### Patch Changes
+
+-   @comet/admin-icons@6.15.1
+
+## 6.15.0
+
+### Patch Changes
+
+-   Updated dependencies [406027806]
+    -   @comet/admin-icons@6.15.0
+
 ## 6.14.1
 
 ### Patch Changes